django/django_project/requirements.txt
psycopg2-binary==2.9.9
Django==5.0.6
python-dotenv==1.0.1
djangorestframework==3.15.1
drf-yasg==1.21.7
setuptools==69.5.1